requirements.txt
requests
beautifulsoup4
lxml
pymongo
dnspython
python-dotenv
joblib